perf(node18): complete updates to <node 18 dependencies
Changes to be committed:
modified:   package-lock.json
modified:   package.json
deleted:    rollup.config.js
new file:   rollup.config.mjs
