+++ package.json
@@ -40,8 +40,7 @@
     "@commitlint/cli": "^17.4.4",
     "@commitlint/config-conventional": "^17.4.4",
     "@open-wc/testing": "^3.1.7",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.6",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@semantic-release/changelog": "^6.0.2",
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/npm": "^9.0.2",
@@ -57,23 +56,21 @@
     "copyfiles": "^2.4.1",
     "core-js": "^3.29.1",
     "eslint": "^8.36.0",
-    "eslint-plugin-jsdoc": "^37.0.3",
+    "eslint-plugin-jsdoc": "^40.1.0",
     "husky": "^8.0.3",
     "lodash": "^4.17.21",
     "markdown-magic": "^2.6.1",
     "nodemon": "^2.0.21",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.21",
-    "postcss-custom-properties": "^12.1.11",
+    "postcss-custom-properties": "^13.1.4",
     "postcss-discard-comments": "^5.1.2",
     "postcss-rem-to-pixel": "^4.1.2",
     "postcss-remove-rules": "^1.0.0",
     "postcss-selector-replace": "^1.0.2",
     "prismjs": "^1.29.0",
-    "rollup": "^2.60.1",
-    "rollup-plugin-minify-html-literals": "^1.2.6",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
+    "rollup": "^3.20.0",
+    "rollup-plugin-serve": "^2.0.2",
     "sass": "^1.59.3",
     "semantic-release": "^20.1.3",
     "sinon": "^15.0.2",
@@ -145,10 +142,9 @@
     "build:ci": "npm-run-all sweep build",
     "build:api": "wca analyze 'src/auro-input.js' --outFiles docs/api.md",
     "build:demo": "npm-run-all build demo:rm:build demo:new:build demo:copy:index demo:copy:demo demo:update:index",
-    "build:dev:assets": "npm-run-all build:sass:demo build:sass:component postCss:component sass:render",
+    "build:dev:assets": "npm-run-all build:sass:component postCss:component sass:render",
     "build:markdownDocs": "node scripts/generateDocs.js",
-    "build:sass": "npm-run-all build:sass:demo build:sass:component postCss:component sass:render",
-    "build:sass:demo": "sass --no-source-map demo:demo",
+    "build:sass": "npm-run-all build:sass:component postCss:component sass:render",
     "build:sass:component": "sass --no-source-map src:src",
     "build:watch": "nodemon -e scss,js --watch src --exec npm run build:dev:assets",
     "bundler": "rollup -c",
